# api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/docs": {
            "get": {
                "summary": "Read Docs",
                "description": "get swagger documentation html",
                "operationId": "read_docs_docs_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "summary": "Create User",
                "description": "create new user with given data",
                "operationId": "create_user_auth_signup_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegistrationInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/update": {
            "post": {
                "summary": "Update User",
                "description": "update user with given data",
                "operationId": "update_user_auth_update_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/signin": {
            "post": {
                "summary": "Create Session",
                "description": "create new session with given data and return secret",
                "operationId": "create_session_auth_signin_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/create": {
            "post": {
                "summary": "Create Post",
                "description": "create new post with given data",
                "operationId": "create_post_post_create_post",
                "parameters": [
                    {
                        "name": "secret",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Secret"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePostInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/update": {
            "post": {
                "summary": "Update Post",
                "description": "update post with given id and data",
                "operationId": "update_post_post_update_post",
                "parameters": [
                    {
                        "name": "secret",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Secret"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePostInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/delete": {
            "delete": {
                "summary": "Delete Post",
                "description": "delete post with given id",
                "operationId": "delete_post_post_delete_delete",
                "parameters": [
                    {
                        "name": "secret",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Secret"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/get": {
            "get": {
                "summary": "Get Post",
                "description": "get post with given id",
                "operationId": "get_post_post_get_get",
                "parameters": [
                    {
                        "name": "secret",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Secret"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/all/get": {
            "get": {
                "summary": "Get All Posts",
                "description": "get all posts with given user and pagination",
                "operationId": "get_all_posts_post_all_get_get",
                "parameters": [
                    {
                        "name": "secret",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Secret"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Id"
                        }
                    },
                    {
                        "name": "start_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Start Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AuthInput": {
                "properties": {
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "password"
                ],
                "title": "AuthInput"
            },
            "CreatePostInput": {
                "properties": {
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "text": {
                        "type": "string",
                        "title": "Text"
                    }
                },
                "type": "object",
                "required": [
                    "title",
                    "text"
                ],
                "title": "CreatePostInput"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "RegistrationInput": {
                "properties": {
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    },
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "second_name": {
                        "type": "string",
                        "title": "Second Name"
                    },
                    "birth_date": {
                        "type": "string",
                        "title": "Birth Date"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "phone_number": {
                        "type": "string",
                        "title": "Phone Number"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "password",
                    "first_name",
                    "second_name",
                    "birth_date",
                    "email",
                    "phone_number"
                ],
                "title": "RegistrationInput"
            },
            "UpdatePostInput": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "text": {
                        "type": "string",
                        "title": "Text"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "title",
                    "text"
                ],
                "title": "UpdatePostInput"
            },
            "UserUpdateInput": {
                "properties": {
                    "secret": {
                        "type": "string",
                        "title": "Secret"
                    },
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "second_name": {
                        "type": "string",
                        "title": "Second Name"
                    },
                    "birth_date": {
                        "type": "string",
                        "title": "Birth Date"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "phone_number": {
                        "type": "string",
                        "title": "Phone Number"
                    }
                },
                "type": "object",
                "required": [
                    "secret",
                    "first_name",
                    "second_name",
                    "birth_date",
                    "email",
                    "phone_number"
                ],
                "title": "UserUpdateInput"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}